# src/main/resources/api.yaml
swagger: "2.0"
info:
  description: "Treditor"
  title: "Treditor"
  version: "1.0.0"
tags:
  - name: treditor
    description: Everything you can edit
paths:
  /projects/{projectName}:
    post:
      produces:
        - application/json
      tags:
        - treditor
      description: create a new project
      operationId: createProject
      parameters:
        - in: path
          name: projectName
          type: string
          required: true
          description: generated project id
      responses:
        '200':
          description: project
          schema:
            $ref: '#/components/schemas/Project'

  /projects/{projectId}:
    get:
      produces:
        - application/json
      tags:
        - treditor
      description: get project with shapes
      operationId: getProject
      parameters:
        - in: path
          name: projectId
          type: string
          required: true
          description: project Id
      responses:
        '200':
            description: project
            schema:
              $ref: '#/components/schemas/Project'
    put:
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - treditor
      operationId: updateProject
      parameters:
        - in: path
          name: projectId
          type: string
          required: true
          description: project Id
        - in: body
          name: projectInfo
          schema:
            $ref: '#/definitions/ProjectInfo'
      responses:
        '200':
          description: пара ключ значение
          schema:
            $ref: '#/components/schemas/Project'
  /projects:
    get:
      produces:
        - application/json
      tags:
        - treditor
      description: get all projects in map id/name
      operationId: getProjects
      responses:
        '200':
          description: пара ключ значение
          schema:
            $ref: '#/definitions/ProjectMap'

definitions:
  ProjectInfo:
    type: object
    description: Информация о проекте
    properties:
      shapes:
        type: array
        description: коллекция всех фигур в проекте
        items:
          $ref: '#/components/schemas/Shape'

  ProjectMap:
    tyoe: object
    additionalProperties:
      #type: '#/components/schemas/ObjectId'
      type: string